fix(security): bind request-derived SQL parameters and rename the sanitizer
pages/clan/detail.php took clanname and clanshort from the request body,
passed them through Sanitize::fullSanitize() and interpolated the result into
an UPDATE inside single quotes. That method removes LotGD colour codes and
escapes nothing, so the values arrived unchanged apart from backtick markup.
The navigation allowlist does not help here: it keys on the request URI, and a
POST body is not part of it. SU_EDIT_COMMENTS is a moderation flag rather than
full superuser, so this was a usable escalation path.

The remaining request-derived interpolations were integer ids that simply
lacked a cast (weapons, mercenarycamp, train, titleedit — the last already
casting in every other use in the same file), plus txnid and a hook-supplied
point total in donators.php, where the neighbouring values were carefully
validated with filter_var and the one gap was easy to miss. clan.php now
normalizes detail at its source because it also reaches a cache key and
generated URLs.

motd.php was exploited through this class of bug before. Its month filter now
binds its values, and the pattern is anchored: previously the length cut and
an unanchored pattern only held together by their interaction, so raising the
limit would have reopened it. The assignment in `if ($newcount = 0)` is fixed
as well; it made the cached branch unreachable.

Sanitize::fullSanitize() becomes stripAllColorCodes(). The old name promised
general purpose sanitization and is how a value came to be treated as safe for
SQL. It stays as a deprecated alias for modules, and the legacy global
full_sanitize() keeps its name and delegates.

SECURITY.md now documents the navigation allowlist as a security control,
including what it does not cover: request bodies, pages that opt out through
OVERRIDE_FORCED_NAV or ALLOW_ANONYMOUS, and values the application itself
interpolates into a nav URL, which widen the allowlist.

+++ mercenarycamp.php
@@ -190,7 +190,7 @@ if ($op == "") {
     Nav::add("Navigation");
     Nav::add("Return to the camp", "mercenarycamp.php?skip=1");
 } elseif ($op == "buy") {
-    $id = Http::get('id');
+    $id = (int) Http::get('id');
     $sql = "SELECT * FROM " . Database::prefix("companions") . " WHERE companionid = $id";
     $result = Database::query($sql);
     if ($row = Database::fetchAssoc($result)) {
